+++ src/main/java/com/pizzaheaven/pizza/repositories/ItemRepository.java
@@ -1,8 +1,7 @@
 package com.pizzaheaven.pizza.repositories;
 
 import com.pizzaheaven.pizza.models.Item;
-import org.hibernate.cache.spi.support.AbstractReadWriteAccess;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
 }
